+++ tests/conftest.py
@@ -1,62 +1,28 @@
-#!/usr/bin/env python
-# -*- coding: utf-8 -*-
+"""Place for fixtures and configuration that will be used in most of the tests.
+A nice option is to put your ``autouse`` fixtures here.
+Functions that can be imported and re-used are more suitable for the ``helpers`` file.
 """
-    Dummy conftest.py for custom_extension.
-
-    If you don't know what this is for, just leave it empty.
-    Read more about conftest.py under:
-    https://pytest.org/latest/plugins.html
-"""
+import logging
 import os
-import shlex
-import stat
-from shutil import rmtree
 
 import pytest
-
 from pyscaffold.exceptions import ShellCommandException
+from pyscaffold.log import ReportFormatter
 
-
-def set_writable(func, path, exc_info):
-    if not os.access(path, os.W_OK):
-        os.chmod(path, stat.S_IWUSR)
-        func(path)
-    else:
-        raise RuntimeError
+from .helpers import rmpath, uniqstr
 
 
 @pytest.fixture
-def tmpfolder(tmpdir):
+def tmpfolder(tmp_path):
     old_path = os.getcwd()
-    newpath = str(tmpdir)
-    os.chdir(newpath)
+    new_path = tmp_path / uniqstr()
+    new_path.mkdir(parents=True, exist_ok=True)
+    os.chdir(str(new_path))
     try:
-        yield tmpdir
+        yield new_path
     finally:
         os.chdir(old_path)
-        rmtree(newpath, onerror=set_writable)
-
-
-@pytest.fixture
-def venv(virtualenv):
-    """Create a virtualenv for each test"""
-    return virtualenv
-
-
-@pytest.fixture
-def venv_run(venv):
-    """Run a command inside the venv"""
-
-    def _run(*args, **kwargs):
-        # pytest-virtualenv doesn't play nicely with external os.chdir
-        # so let's be explicit about it...
-        kwargs["cd"] = os.getcwd()
-        kwargs["capture"] = True
-        if len(args) == 1 and isinstance(args[0], str):
-            args = shlex.split(args[0])
-        return venv.run(args, **kwargs).strip()
-
-    return _run
+        rmpath(new_path)
 
 
 @pytest.fixture
@@ -66,3 +32,34 @@ def nodjango_admin_mock(monkeypatch):
 
     monkeypatch.setattr("pyscaffoldext.django.extension.django_admin", raise_error)
     yield
+
+
+@pytest.fixture(autouse=True)
+def isolated_logger(request, monkeypatch):
+    """See isolated_logger in pyscaffold/tests/conftest.py to see why this fixture
+    is important to guarantee tests checking logs work as expected.
+    This just work for multiprocess environments, not multithread.
+    """
+    if "original_logger" in request.keywords:
+        yield
+        return
+
+    # Get a fresh new logger, not used anywhere
+    raw_logger = logging.getLogger(uniqstr())
+    raw_logger.setLevel(logging.NOTSET)
+    new_handler = logging.StreamHandler()
+
+    patches = {
+        "propagate": True,  # <- needed for caplog
+        "nesting": 0,
+        "wrapped": raw_logger,
+        "handler": new_handler,
+        "formatter": ReportFormatter(),
+    }
+    for key, value in patches.items():
+        monkeypatch.setattr(f"pyscaffold.log.logger.{key}", value)
+
+    try:
+        yield
+    finally:
+        new_handler.close()