+++ CHANGELOG.rst
@@ -2,7 +2,10 @@
 Changelog
 =========
 
-Version 0.1
-===========
+Version 0.1 (development)
+=========================
 
-- Initial release
+- Initial release extracted from PyScaffold
+- Updated to PyScaffold v4
+- Added ``__main__.py`` and custom ``manage.py`` to bypass Django's packaging limitations, PR #2
+- Added support for project path different from package name